LayoutTests/http/tests/w3c/html/semantics/embedded-content/the-img-element/srcset/parse-a-srcset-attribute-expected.txt

                                                                                                                                                                                                                                             

FAIL "" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "," assert_equals: expected (string) "" but got (undefined) undefined
FAIL ",,," assert_equals: expected (string) "" but got (undefined) undefined
FAIL "  data:,a  1x  " assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "\t\tdata:,a\t\t1x\t\t" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "\n\ndata:,a\n\n1x\n\n" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "\v\vdata:,a\v\v1x\v\v" assert_equals: expected (string) "data:,a%0B%0B1x" but got (undefined) undefined
FAIL "\f\fdata:,a\f\f1x\f\f" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "\r\rdata:,a\r\r1x\r\r" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "\x0e\x0edata:,a\x0e\x0e1x\x0e\x0e" assert_equals: expected (string) "data:,a%0E%0E1x" but got (undefined) undefined
FAIL "\x0f\x0fdata:,a\x0f\x0f1x\x0f\x0f" assert_equals: expected (string) "data:,a%0F%0F1x" but got (undefined) undefined
FAIL "\x10\x10data:,a\x10\x101x\x10\x10" assert_equals: expected (string) "data:,a%10%101x" but got (undefined) undefined
FAIL "data:,a" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a " assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a ," assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a," assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a, " assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a,,," assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a,, , " assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL " data:,a" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL ",,,data:,a" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL " , ,,data:,a" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL " data:,a" assert_equals: expected (string) "http://127.0.0.1:8000/w3c/html/semantics/embedded-content/the-img-element/srcset/%C2%A0data:,a" but got (undefined) undefined
FAIL "data:,a " assert_equals: expected (string) "data:,a%C2%A0" but got (undefined) undefined
FAIL "data:,a 1x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1x " assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1x," assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a ( , data:,b 1x, ), data:,c" assert_equals: expected (string) "data:,c" but got (undefined) undefined
FAIL "data:,a ((( , data:,b 1x, ), data:,c" assert_equals: expected (string) "data:,c" but got (undefined) undefined
FAIL "data:,a [ , data:,b 1x, ], data:,c" assert_equals: expected (string) "data:,b" but got (undefined) undefined
FAIL "data:,a { , data:,b 1x, }, data:,c" assert_equals: expected (string) "data:,b" but got (undefined) undefined
FAIL "data:,a \" , data:,b 1x, \", data:,c" assert_equals: expected (string) "data:,b" but got (undefined) undefined
FAIL "data:,a \\,data:;\,b, data:,c" assert_equals: expected (string) "data:;\\,b" but got (undefined) undefined
FAIL "data:,a, data:,b (" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a, data:,b (  " assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a, data:,b (," assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a, data:,b (x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a, data:,b ()" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a (, data:,b" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a /*, data:,b, data:,c */" assert_equals: expected (string) "data:,b" but got (undefined) undefined
FAIL "data:,a //, data:,b" assert_equals: expected (string) "data:,b" but got (undefined) undefined
FAIL "data:,a foo" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a foo foo" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a foo 1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a foo 1x foo" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a foo 1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a foo 1w foo" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1x 1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1x 1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1h" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1h 1w" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1h 1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1h 1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1h 1w 1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1x 1w 1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1h foo" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a foo 1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a -1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w -1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1.0w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1.0w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1e0w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1e0w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1www" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1www" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a +1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w +1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1W" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1W" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a Infinityw" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w Infinityw" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a NaNw" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w NaNw" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0x1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0X1w" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1\x01w" (trailing U+0001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+00A0) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+1680) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2002) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2003) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2004) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2005) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2006) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2007) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2008) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+2009) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+200A) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1‌w" (trailing U+200C) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1‍w" (trailing U+200D) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+202F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 w" (trailing U+205F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1　w" (trailing U+3000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1﻿w" (trailing U+FEFF) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a \x011w" (leading U+0001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+00A0) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+1680) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2002) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2003) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2004) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2005) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2006) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2007) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2008) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+2009) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+200A) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a ‌1w" (leading U+200C) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a ‍1w" (leading U+200D) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+202F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1w" (leading U+205F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 　1w" (leading U+3000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a ﻿1w" (leading U+FEFF) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a -0x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1x -0x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a -1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1x -1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1e0x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1E0x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1e-1x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1.5e1x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a -x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a .x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a -.x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1.x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a .5x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a .5e1x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1x 1.5e1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1x 1e1.5x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1.0x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1x 1.0x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a +1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1X" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a Infinityx" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a NaNx" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0x1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0X1x" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1\x01x" (trailing U+0001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+00A0) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+1680) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2002) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2003) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2004) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2005) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2006) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2007) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2008) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+2009) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+200A) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1‌x" (trailing U+200C) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1‍x" (trailing U+200D) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+202F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1 x" (trailing U+205F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1　x" (trailing U+3000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1﻿x" (trailing U+FEFF) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a \x011x" (leading U+0001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+00A0) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+1680) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2002) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2003) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2004) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2005) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2006) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2007) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2008) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+2009) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+200A) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a ‌1x" (leading U+200C) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a ‍1x" (leading U+200D) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+202F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a  1x" (leading U+205F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 　1x" (leading U+3000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a ﻿1x" (leading U+FEFF) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 0h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w -1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1.0h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1e0h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1hhh" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w +1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1H" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w Infinityh" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w NaNh" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0x1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 0X1h" assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1\x01h" (trailing U+0001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+00A0) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+1680) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2002) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2003) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2004) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2005) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2006) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2007) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2008) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+2009) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+200A) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1‌h" (trailing U+200C) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1‍h" (trailing U+200D) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+202F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1 h" (trailing U+205F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1　h" (trailing U+3000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 1﻿h" (trailing U+FEFF) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w \x011h" (leading U+0001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+00A0) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+1680) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2001) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2002) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2003) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2004) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2005) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2006) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2007) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2008) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+2009) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+200A) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w ‌1h" (leading U+200C) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w ‍1h" (leading U+200D) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+202F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w  1h" (leading U+205F) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w 　1h" (leading U+3000) assert_equals: expected (string) "" but got (undefined) undefined
FAIL "data:,a 1w ﻿1h" (leading U+FEFF) assert_equals: expected (string) "" but got (undefined) undefined
